feat: Cleaning up the /dist folder
As you might have noticed over the past guides and code example, our /dist folder has become quite cluttered. Webpack will generate the files and put them in the /dist folder for you, but it doesn't keep track of which files are actually in use by your project.

In general it's good practice to clean the /dist folder before each build, so that only used files will be generated. Let's take care of that.

A popular plugin to manage this is the clean-webpack-plugin so let's install and configure it.

npm install clean-webpack-plugin --save-dev

Now run an npm run build and inspect the /dist folder. If everything went well you should now only see the files generated from the build and no more old files.